refactor: restructure app module
