+++ packages/app/src/index.ts
@@ -1,4 +1,5 @@
-export * from "./public-types";
-export * from "./api";
-export * from "./errors";
+export * from "./types";
+export * from "./shared/errors";
 export { default as Logger } from "./lib/Logger";
+export { default as initializeApp, type NavigraphApp } from "./lib/initializeApp";
+export { getApp } from "./internals/apps";
