--- packages/app/src/internal.ts
@@ -1,3 +1,0 @@
-import { NavigraphApp } from "./public-types";
-
-export const _apps = new Map<string, NavigraphApp>();
