+++ packages/app/src/shared/errors.ts
