+++ packages/app/src/types.ts
@@ -1,0 +1,8 @@
+export enum Scope {
+  "CHARTS" = "charts",
+  "FMSDATA" = "fmsdata",
+  "EMAIL" = "email",
+  "TILES" = "tiles",
+  /** @deprecated Included by default, no need to include it explicitly */
+  "OFFLINE" = "offline_access",
+}
